+++ puffin-imgui/examples/imgui.rs
@@ -149,7 +149,7 @@ fn main() {
             std::thread::sleep(std::time::Duration::from_millis(10))
         }
 
-        for _ in 0..=1000 {
+        for _ in 0..1000 {
             puffin::profile_scope!("very thin");
         }